+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNP1EM6TVJ3J5FF3R3S3H7DT"
+        "01KNP589GVSAPHZWDV9X5TBZCH"
     )
     r"""The unique identifier of the data source"""