+++ packages/orq-rc/RELEASES.md
@@ -6919,3 +6919,13 @@ Based on:
 - [python v4.7.0-rc.52] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.52] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.52 - packages/orq-rc
+
+## 2026-04-08 09:04:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.53] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.53 - packages/orq-rc